+++ cosmos/operators/_watcher/__init__.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
-__all__ = ["get_xcom_val", "safe_xcom_push", "build_producer_state_fetcher"]
+__all__ = ["get_xcom_val", "safe_xcom_push", "build_producer_state_fetcher", "WatcherTrigger", "_parse_compressed_xcom"]
 
 from cosmos.operators._watcher.state import build_producer_state_fetcher, get_xcom_val, safe_xcom_push
+from cosmos.operators._watcher.triggerer import WatcherTrigger, _parse_compressed_xcom
